Add 'PseudoTerminal' to 'TestSupport'
