+++ Tests/BasicTests/TerminalControllerTests.swift
@@ -10,40 +10,9 @@
 
 import XCTest
 @testable import Basic
+@testable import TestSupport
 import SPMLibc
 
-final class PseudoTerminal {
-    let master: Int32
-    let slave: Int32
-    var outStream: LocalFileOutputByteStream
-
-    init?(){
-        var master: Int32 = 0
-        var slave: Int32 = 0
-        if openpty(&master, &slave, nil, nil, nil) != 0 {
-            return nil
-        }
-        guard let outStream = try? LocalFileOutputByteStream(filePointer: fdopen(slave, "w"), closeOnDeinit: false) else {
-            return nil
-        }
-        self.outStream = outStream
-        self.master = master
-        self.slave = slave
-    }
-
-    func readMaster(maxChars n: Int = 1000) -> String? {
-        var buf: [CChar] = [CChar](repeating: 0, count: n)
-        if read(master, &buf, n) <= 0 {
-            return nil
-        }
-        return String(cString: buf)
-    }
-
-    func close() {
-        _ = SPMLibc.close(slave)
-        _ = SPMLibc.close(master)
-    }
-}
 
 final class TerminalControllerTests: XCTestCase {
     func testBasic() {
