+++ Tests/UtilityTests/ProgressBarTests.swift
@@ -12,45 +12,9 @@ import XCTest
 import Utility
 import SPMLibc
 @testable import Basic
+@testable import TestSupport
 
 typealias Thread = Basic.Thread
-
-// FIXME: Copied from BasicTests, move to TestSupport once available.
-final class PseudoTerminal {
-    let master: Int32
-    let slave: Int32
-    var outStream: LocalFileOutputByteStream
-
-    init?(){
-        var master: Int32 = 0
-        var slave: Int32 = 0
-        if openpty(&master, &slave, nil, nil, nil) != 0 {
-            return nil
-        }
-        guard let outStream = try? LocalFileOutputByteStream(filePointer: fdopen(slave, "w"), closeOnDeinit: false) else {
-            return nil
-        }
-        self.outStream = outStream
-        self.master = master
-        self.slave = slave
-    }
-
-    func readMaster(maxChars n: Int = 1000) -> String? {
-        var buf: [CChar] = [CChar](repeating: 0, count: n)
-        if read(master, &buf, n) <= 0 {
-            return nil
-        }
-        return String(cString: buf)
-    }
-
-    func closeSlave() {
-        _ = SPMLibc.close(slave)
-    }
-
-    func closeMaster() {
-        _ = SPMLibc.close(master)
-    }
-}
 
 final class ProgressBarTests: XCTestCase {
     func testProgressBar() {
@@ -88,10 +52,10 @@ final class ProgressBarTests: XCTestCase {
         }
         thread.start()
         runProgressBar(bar)
-        pty.closeSlave()
         // Make sure to read the complete output before checking it.
+        pty.close()
         thread.join()
-        pty.closeMaster()
+        pty.close()
         XCTAssertTrue(output.chuzzle()?.hasPrefix("\u{1B}[36m\u{1B}[1mTestHeader\u{1B}[0m") ?? false)
     }
